refactor: created a seperate routes file for destinations. updated that in backend and frontend
