+++ backend/src/presentation/constants/routes/baseRoutes.ts
@@ -4,4 +4,5 @@ export enum Routes {
   CAB = "/api/v1/cab",
   HOTEL = "/api/v1/hotels",
   ROOM = "/api/v1/rooms",
+  DESTINATION = "/api/v1/destinations",
 }
