+++ backend/src/presentation/constants/routes/destinationRoutes.ts
@@ -1,0 +1,4 @@
+export enum DESTINATION_ROUTES {
+  INDEX = "/",
+  DESTINATION_IMAGE = "/:id/images/:index",
+}
