+++ frontend/src/constants/routes/adminRoutes.ts
@@ -4,6 +4,4 @@ export enum AdminRoutes {
     UNVERIFIED_USERS = "/admin/users/verification",
     REJECT_USER = "/admin/users/verification/##id##/reject",
     APPROVE_USER = "/admin/users/verification/##id##/approve",
-    DESTINATIONS = "/admin/destinations",
-    DESTINATION_IMAGE = "/admin/destinations/##id##/images/##index##",
 }
