+++ frontend/src/constants/routes/destinationRoutes.ts
@@ -1,0 +1,4 @@
+export enum DESTINATION_ROUTES {
+    DESTINATIONS = "/destinations",
+    DESTINATION_IMAGE = "/destinations/##id##/images/##index##",
+}
