+++ frontend/src/constants/routes/roomRoutes.ts
@@ -1,5 +1,4 @@
 export enum ROOM_ROUTES {
-    CREATE = "/rooms/hotel/:hotelId",
     BY_HOTEL = "/rooms/hotel/:hotelId",
     BY_ID = "/rooms/:id",
 }
