+++ frontend/src/queryOptions/adminQueryOptions.ts
@@ -2,26 +2,18 @@
 import { QUERY_KEYS } from "@/constants/queryKeys/queryKeys";
 import type { filterType } from "@/pages/admin/Users";
 import {
-    addDestination,
     approveUserVerificationRequest,
     changeUserStatus,
-    DeleteDestinationImage,
-    findDestinationById,
-    findDestinations,
     getUnverifiedUserDetails,
     getUnverifiedUsers,
     getUsers,
     rejectUserVerificationRequest,
-    updateDestination,
 } from "@/services/api/adminApiService";
 import type {
-    IFindDestinationsRequest,
     IRejectUserRegistrationRequest,
     IUserStatusChangeRequest,
 } from "@/types/api/requests/adminRequest";
 import type { IGetUsersResponse } from "@/types/api/responses/adminResponse";
-import type { IFindDestinationsResponse } from "@/types/api/responses/findDestinationAdminResponse";
-import type { IFindDestinationResponseDTO } from "@/types/api/responses/findDestinationResponse";
 import type { IGetUnverifiedUserDetailsResponseDTO } from "@/types/api/responses/getUnverifiedUserDetailsResponse";
 import type { IGetUnverifiedUsersResponseDTO } from "@/types/api/responses/getUnverifiedUsersResponse";
 import type { HttpResponse } from "@/types/api/responseType";
@@ -85,55 +77,5 @@ export function createRejectUserVerificationRequestMutationOption(id: string) {
 export function createApproveUserVerificationRequestMutationOption(id: string) {
     return mutationOptions<HttpResponse<object>, Error>({
         mutationFn: () => approveUserVerificationRequest(id),
-    });
-}
-
-export function createAddDestinationMutationOption() {
-    return mutationOptions<HttpResponse<object>, Error, FormData>({
-        mutationFn: addDestination,
-    });
-}
-
-export function createFindDestinationsQueryOption(
-    filter: IFindDestinationsRequest
-) {
-    return queryOptions<HttpResponse<IFindDestinationsResponse>, Error>({
-        queryKey: [
-            QUERY_KEYS.DESTINATIONS,
-            filter.pageNo,
-            filter.query,
-            filter.type,
-            filter.isFree,
-            filter.isActive,
-        ],
-        queryFn: () => findDestinations(filter),
-        placeholderData: keepPreviousData,
-    });
-}
-
-export function createUpdateDestinationMutationOption() {
-    return mutationOptions<
-        HttpResponse<object>,
-        Error,
-        { id: string; data: FormData }
-    >({
-        mutationFn: updateDestination,
-    });
-}
-
-export function createFindDestinationByIdQueryOption(id: string) {
-    return queryOptions<HttpResponse<IFindDestinationResponseDTO>, Error>({
-        queryKey: [QUERY_KEYS.DESTINATIONS, id],
-        queryFn: () => findDestinationById(id),
-    });
-}
-
-export function createDeleteDestinationImageMutationOption() {
-    return mutationOptions<
-        HttpResponse<object>,
-        Error,
-        { id: string; index: number }
-    >({
-        mutationFn: (data) => DeleteDestinationImage(data.id, data.index),
     });
 }
